[move] domain.minecraft.UuidRepositoryをsubsystemに移す
